// src/types/database.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
// ...
  public: {
    Tables: {
// ...
      venues: {
        Row: {
          address_district: string
          address_full: string
          amenities: string[]
          created_at: string
          id: string
          location: unknown
          name: string
          operating_hours: Json
          parking_available: boolean
          phone: string | null
          total_seats: number | null
          updated_at: string
        }
        Insert: {
          address_district: string
          address_full: string
          amenities?: string[]
          created_at?: string
          id?: string
          location: unknown
          name: string
          operating_hours?: Json
          parking_available?: boolean
          phone?: string | null
          total_seats?: number | null
          updated_at?: string
        }
        Update: {
          address_district?: string
          address_full?: string
          amenities?: string[]
          created_at?: string
          id?: string
          location?: unknown
          name?: string
          operating_hours?: Json
          parking_available?: boolean
          phone?: string | null
          total_seats?: number | null
          updated_at?: string
        }
        Relationships: []
      }
    }
// ...
  }
}